fix voites
