+++ config.py
@@ -1,18 +1,10 @@
 from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
+
 
 import platform
 import secrets
 
 app = FastAPI()
-
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
 
 SECRET_KEY = secrets.token_urlsafe(32)
 ALGORITHM = "HS256"
@@ -22,5 +14,5 @@ class Config:
     DB_USER = "postgres"
     DB_PASS = "postgres"
     DB_NAME = "dgtu"
-    DB_HOST = "db"  # Используйте имя сервиса Docker в качестве хоста
+    DB_HOST = "localhost" 
     DB_PORT = "5432"
